backend/pyproject.toml
[project]
name = "pgvector-workbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Santosh Kanumuri", email = "dkanumur@ttu.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "asyncpg==0.29.0",
    "bcrypt>=4.0.0",
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "passlib==1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose==3.3.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "scikit-learn>=1.7.2",
    "umap-learn>=0.5.9.post2",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
pgvector-workbench = "pgvector_workbench:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
